refactor: 파일 위치 변경에 따른 주석 변경
파일 위치 변경에 따른 주석 변경사항입니다.
