+++ src/main/java/mana/doodleking/domain/room/dto/RoomDetail.java
@@ -1,7 +1,7 @@
 package mana.doodleking.domain.room.dto;
 
 import lombok.*;
-import mana.doodleking.domain.room.Room;
+import mana.doodleking.domain.room.domain.Room;
 import mana.doodleking.domain.room.enums.RoomState;
 import mana.doodleking.domain.room.enums.Subject;
