+++ src/main/java/mana/doodleking/domain/room/repository/RoomRepository.java
@@ -1,6 +1,6 @@
 package mana.doodleking.domain.room.repository;
 
-import mana.doodleking.domain.room.Room;
+import mana.doodleking.domain.room.domain.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
