+++ src/main/java/mana/doodleking/domain/room/repository/UserRoomRepository.java
@@ -1,6 +1,6 @@
 package mana.doodleking.domain.room.repository;
 
-import mana.doodleking.domain.room.UserRoom;
+import mana.doodleking.domain.room.domain.UserRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
