+++ src/main/java/mana/doodleking/domain/user/repository/CharacterRepository.java
@@ -2,7 +2,7 @@ package mana.doodleking.domain.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import mana.doodleking.domain.user.Character;
+import mana.doodleking.domain.user.domain.Character;
 
 @Repository
 public interface CharacterRepository extends JpaRepository<Character, Long> {
