+++ src/main/java/mana/doodleking/domain/user/service/UserService.java
@@ -1,8 +1,8 @@
 package mana.doodleking.domain.user.service;
 
 import lombok.RequiredArgsConstructor;
-import mana.doodleking.domain.user.Character;
-import mana.doodleking.domain.user.User;
+import mana.doodleking.domain.user.domain.Character;
+import mana.doodleking.domain.user.domain.User;
 import mana.doodleking.domain.user.dto.CreateUserReq;
 import mana.doodleking.domain.user.repository.CharacterRepository;
 import mana.doodleking.domain.user.repository.UserRepository;
