+++ src/test/java/mana/doodleking/domain/user/service/UserServiceTest.java
@@ -1,7 +1,7 @@
 package mana.doodleking.domain.user.service;
 
-import mana.doodleking.domain.user.Character;
-import mana.doodleking.domain.user.User;
+import mana.doodleking.domain.user.domain.Character;
+import mana.doodleking.domain.user.domain.User;
 import mana.doodleking.domain.user.dto.CreateUserReq;
 import mana.doodleking.domain.user.repository.CharacterRepository;
 import mana.doodleking.domain.user.repository.UserRepository;
